feat: Scroll article card to 1/3 of a view in landscape mode
To allow the user to see whole background image, add extra top padding to scroll view so only 24dp of the card are visible on the bottom of the screen when fully scrolled down. When user opens the article the card will be automatically scrolled up to 1/3 of the screen.