fix: Resolve detail activity crash
The app would crash sometimes when detail activity was launched or page has been changed. It was caused by an attempt to set up scroll view before fragment was attached to the activity. To resolve the issue a condition has been added to ensure troublesome code is executed only when fragment is displayed.