feat: Increase ViewPager offscreen page limit
With default offscreen page limit (2), new pages would pop out on page transformation instead of fade in. It happened because new fragment was created with alpha set to 1. Then PageTransformer would change alpha to 0 and fade in. By changing offscreen page limit to 4 we ensure that next page is always ready by the time we reach it.